refactor: refactor endpoints
