+++ .gitignore
@@ -37,5 +37,5 @@ venv/
 ENV/
 env.bak/
 venv.bak/
-secrets/*
+secrets/
 # End of https://www.toptal.com/developers/gitignore/api/python
