+++ pyproject.toml
@@ -10,8 +10,9 @@ fastapi = "^0.70.0"
 uvicorn = "^0.15.0"
 SQLAlchemy = "^1.4.27"
 alembic = "^1.7.5"
-psycopg2-binary = "^2.9.2"
 python-multipart = "^0.0.5"
+psycopg2 = "^2.9.2"
+postgres = "^4.0"
 
 [tool.poetry.dev-dependencies]
 flake8 = "^4.0.1"
