--- secrets/.env.dev
@@ -1,4 +1,0 @@
-DB_USER = user
-DB_PASSWORD = password
-DB_HOST = db
-DB_NAME = per4mance
