+++ src/course/endpoints.py
@@ -2,44 +2,20 @@ from typing import Optional
 
 from fastapi.param_functions import Depends, Path, Query
 from fastapi.routing import APIRouter
-from fastapi.security.oauth2 import OAuth2PasswordRequestForm
 from starlette.responses import JSONResponse
 
-from per4mance.course.schemas import CoursePostSchema
-from per4mance.course.services import (
+from src.course.schemas import CoursePostSchema
+from src.course.services import (
     create_course,
     create_coursexuser,
     delete_course,
     fetch_courses,
     update_course,
 )
-from per4mance.models import User
-from per4mance.user.auth import get_current_user
-from per4mance.user.schemas import SignUp
-from per4mance.user.services import create_account, refresh_token
+from src.models import User
+from src.core.auth import get_current_user
 
 router = APIRouter()
-
-
-@router.post("/users/signup")
-async def signup(signup_form: SignUp) -> JSONResponse:
-    """
-    Create account
-    """
-    token = await create_account(signup_form)
-    return JSONResponse(status_code=201, content=dict(token=token))
-
-
-@router.post("/users/login")
-async def login(signin_form: OAuth2PasswordRequestForm = Depends()) -> JSONResponse:
-    """
-    Get fresh token about requested user info.
-    """
-    token = await refresh_token(signin_form)
-    return JSONResponse(
-        status_code=200,
-        content=dict(access_token=token, token_type="bearer"),
-    )
 
 
 @router.get("/course")
@@ -75,7 +51,6 @@ async def patch_course(
     """
     Update course info.
     """
-    print(course_form)
     course = await update_course(
         course_id=course_id, course_info=course_form, user=evaluator
     )
