+++ src/course/services.py
@@ -4,10 +4,10 @@ from typing import Any, Dict, List, Tuple
 import sqlalchemy as sa
 from fastapi.exceptions import HTTPException
 
-from per4mance import db_engine
-from per4mance.core.utils import fetch_all
-from per4mance.course.schemas import CoursePostSchema
-from per4mance.models import Course, CourseXStudent, User
+from src import db_engine
+from src.core.utils import fetch_all
+from src.course.schemas import CoursePostSchema
+from src.models import Course, CourseXStudent, User
 
 
 async def fetch_courses(limit: int, offset: int, user: User) -> None:
