+++ src/user/endpoints.py
@@ -1,0 +1,31 @@
+from fastapi.security.oauth2 import OAuth2PasswordRequestForm
+
+from starlette.responses import JSONResponse
+from fastapi.routing import APIRouter
+from src.user.schemas import SignUp
+from src.user.services import create_account, refresh_token
+
+from fastapi.param_functions import Depends
+
+router = APIRouter()
+
+
+@router.post("/users/signup")
+async def signup(signup_form: SignUp) -> JSONResponse:
+    """
+    Create account
+    """
+    token = await create_account(signup_form)
+    return JSONResponse(status_code=201, content=dict(token=token))
+
+
+@router.post("/users/login")
+async def login(signin_form: OAuth2PasswordRequestForm = Depends()) -> JSONResponse:
+    """
+    Get fresh token about requested user info.
+    """
+    token = await refresh_token(signin_form)
+    return JSONResponse(
+        status_code=200,
+        content=dict(access_token=token, token_type="bearer"),
+    )
